Dynamic version and requirements driven dependencies
We deduplicate the package dependencies by having the packaing tool
read them from the requirements folder. For versioning, source
control can be used to link versions to tags.

The more work that can be done automatically the more reliable its
use will be. More information in the readme.

This is complemented by a github workflow that will package the
code when a tag is uploaded to origin.